==> 2-Data-Cleaning-NLP.ipynb <==
import numpy as np
import pandas as pd
from collections import defaultdict

df_2019 = pd.read_pickle("./df_2019.pkl")

df_2019.head(3)

from langdetect import detect

# Reduce to only English tweets:
df_2019['lang'] = df_2019['tweet'].apply(lambda row: detect(row))
df_2019 = df_2019[df_2019['lang']=='en']
df_2019.head(3)

## Text Cleaning:

### Option 1: Clean with regex and fit with NLPpipe

from sklearn.feature_extraction.text import CountVectorizer
from langdetect import detect
import pickle
from NLPPipe import NLPPipe

import re
import string

# Look for all unicode emoji ranges:
def deEmojify(text):
    regrex_pattern = re.compile(pattern = "["
        u"\U0001F600-\U0001F64F"  # emoticons
        u"\U0001F300-\U0001F5FF"  # symbols & pictographs
        u"\U0001F680-\U0001F6FF"  # transport & map symbols
        u"\U0001F1E0-\U0001F1FF"  # flags (iOS)
                           "]+", flags = re.UNICODE)
    return regrex_pattern.sub(r'',text)

# Set other cleaning lambdas, in order:
breaks = lambda x: re.sub("\n", " ", x)
links = lambda x: re.sub(r"http\S+", "", x)
hashtag = lambda x: re.sub(r'\B#\w*[a-zA-Z]+\w*','', x)
alphanumeric = lambda x: re.sub("\w*\d\w*", "", x)
punc_lower = lambda x: re.sub("[%s]" % re.escape(string.punctuation), "", x.lower())

df_2019["cleaned_tweet"] = (df_2019["tweet"].apply(breaks)
                                            .apply(links)
                                            .apply(hashtag)
                                            .apply(alphanumeric)
                                            .apply(punc_lower)
                                            .apply(deEmojify)
                            )
df_2019.head(3)

df_2019.to_pickle("./2019_cleaned.pkl")

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer, ENGLISH_STOP_WORDS
from nltk.tokenize import TreebankWordTokenizer, TweetTokenizer
from nltk.stem import PorterStemmer, WordNetLemmatizer
from nltk import word_tokenize, FreqDist
from nltk.corpus import stopwords
from sklearn.decomposition import NMF

def clean_text(text, tokenizer, stemmer):    
    cleaned_text = []
    for post in text:
        post = re.sub("\n", " ", post) # remove newlines
        post = re.sub(r"http\S+", "", post) # remove links
        post = re.sub(r'\B#\w*[a-zA-Z]+\w*','', post) # remove hashtags
        post = re.sub("\w*\d\w*", "", post)
        post = re.sub("[%s]" % re.escape(string.punctuation), "", post.lower())
        cleaned_words = []
        for word in tokenizer(post):
            low_word = word.lower()
            if stemmer:
                low_word = stemmer.stem(low_word)
            cleaned_words.append(low_word)
        cleaned_text.append(' '.join(cleaned_words))
    return cleaned_text

nlp = NLPPipe(vectorizer=CountVectorizer(), 
              cleaning_function=clean_text,
              tokenizer=TreebankWordTokenizer().tokenize, 
              stemmer=PorterStemmer())

corpus = df_2019['tweet'].tolist()
nlp.fit(corpus)
nlp.transform(corpus).toarray()

nlp.vectorizer.vocabulary_

I thought about lemmatizing here, but we've already tokenized, so let's move on.

stop_words = ENGLISH_STOP_WORDS.union(['neruips','neurips2019'])
tfidf = TfidfVectorizer(stop_words=stop_words)
tweet_word_matrix = tfidf.fit_transform(df_2019['cleaned_tweet'])

vocab = tfidf.get_feature_names()
vocab

nlp.__dict__.keys()

### NMF

nmf = NMF(n_components=5)
nmf.fit(tweet_word_matrix)

#### Tweet/Topic Matrix

tweet_topic_matrix = nmf.transform(tweet_word_matrix)
tweet_topic_matrix_df = pd.DataFrame(tweet_topic_matrix).add_prefix('topic_')

tweet_topic_matrix_df[['raw_tweets', 'clean_tweets']] = df_2019[['tweet', 'cleaned_tweet']]
tweet_topic_matrix_df.head()

tweet_topic_matrix_df.to_pickle("./2019_tweet_topic.pkl")

#### Word/Topic Matrix

word_topic_matrix_df = pd.DataFrame(nmf.components_, columns=vocab).T.add_prefix('topic_')
word_topic_matrix_df.head()

## Topic Interpretation

for tweet in tweet_topic_matrix_df.sort_values(by='topic_0', ascending=False).head(10)['raw_tweets'].values:
    print(tweet)
    print()

word_topic_matrix_df.sort_values(by='topic_0', ascending=False).head(10)

def top_tweets(tweet_topic_matrix_df, topic, n_tweets):
    return (tweet_topic_matrix_df
            .sort_values(by=topic, ascending=False)
            .head(n_tweets)['raw_tweets']
            .values)

def top_words(word_topic_matrix_df, topic, n_words):
    return (word_topic_matrix_df
            .sort_values(by=topic, ascending=False)
            .head(n_words))[topic]

for tweet in top_tweets(tweet_topic_matrix_df, 'topic_1', 5):
    print(tweet)
    print()

Topic 1 seems to refer to Yoshua Bengio's focus on Reinforcement Learning.

top_words(word_topic_matrix_df, 'topic_1', 10)

for tweet in top_tweets(tweet_topic_matrix_df, 'topic_2', 5):
    print(tweet)
    print()

top_words(word_topic_matrix_df, 'topic_2', 10)

This topic is a discussion of the mathematical rigor of NeurIPS submissions.

### Option 2: Clean with spaCy

import spacy
nlp = spacy.load('en_core_web_sm')

spacy_tweet = df_2019.iloc[10]['tweet']
doc = nlp(spacy_tweet)

for token in doc:
    print(token.text, token.pos_, token.lemma_, token.is_stop)

doc.text

for sent in doc.sents:
    print(sent)

from spacy import displacy

displacy.render(doc, style='dep', options={'distance': 80})

df_2019['spacy_tweet'] = list(nlp.pipe(df_2019.tweet))

adj = [token.text.lower() for doc in df_2019.spacy_tweet for token in doc if token.pos_=='ADJ']
noun = [token.text.lower() for doc in df_2019.spacy_tweet for token in doc if token.pos_=='NOUN']

from collections import Counter

Counter(adj).most_common(10)

Counter(noun).most_common(10)

spacy.displacy.render(doc, style='ent')
