==> 5-Embeddings.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

from sklearn.cluster import KMeans 
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer, ENGLISH_STOP_WORDS
from sklearn.decomposition import PCA
from sklearn.preprocessing import normalize
from sklearn.metrics import pairwise_distances

import nltk
import string

df = pd.read_pickle("./2019_cleaned.pkl")

df.head()

corpus = df['cleaned_tweet']

vectorizer = CountVectorizer(ngram_range=(1,3))
X = vectorizer.fit_transform(corpus)

print(f"Dimensions of Document-term matrix: {X.toarray().shape}")

vectorizer = CountVectorizer(ngram_range=(1,3), stop_words='english')
X = vectorizer.fit_transform(corpus)

print(f"Dimensions of Document-term matrix: {X.toarray().shape}")

import gensim

# Tokenize the documents
tokenized_docs = [gensim.utils.simple_preprocess(d) for d in corpus]

# Create a Gensim Dictionary.  This creates an id to word mapping for everything in our vocbulary
# It is NOT the same as the dictionary object in the Python standard library
mydict = gensim.corpora.Dictionary()

# Create a Gensim Corpus object.  This creates a list of tuples for each document.
# The first element of the tuple is the word id, the second is the number of counts
mycorpus = [mydict.doc2bow(doc, allow_update=True) for doc in tokenized_docs]

doc_term_matrix = gensim.matutils.corpus2dense(mycorpus, num_terms=len(mydict))

doc_term_matrix
tfidf = gensim.models.TfidfModel(mycorpus)
tfidf_matrix = gensim.matutils.corpus2dense(tfidf[mycorpus], num_terms=len(mydict))


tfidf_matrix

tokenized_docs = [gensim.utils.simple_preprocess(d) for d in corpus]

model = gensim.models.Word2Vec(tokenized_docs, size=10, window=2,min_count=1, sg=1)

import gensim.downloader as api
from gensim.test.utils import datapath

google_vec_file = './GoogleNews-vectors-negative300.bin'
model = api.load('word2vec-google-news-300')
